Module refactoring part 1
