+++ gneiss-mqtt/src/mqtt/auth.rs
@@ -4,7 +4,7 @@
  */
 
 
-use crate::decode::utils::*;
+use crate::decode::*;
 use crate::encode::*;
 use crate::encode::utils::*;
 use crate::error::{MqttError, MqttResult};
