+++ gneiss-mqtt/src/mqtt/connect.rs
@@ -4,7 +4,7 @@
  */
 
 #[cfg(test)]
-use crate::decode::utils::*;
+use crate::decode::*;
 use crate::encode::*;
 use crate::encode::utils::*;
 use crate::error::{MqttError, MqttResult};
