+++ gneiss-mqtt/src/mqtt/publish.rs
@@ -4,7 +4,7 @@
  */
 
 use crate::alias::*;
-use crate::decode::utils::*;
+use crate::decode::*;
 use crate::encode::*;
 use crate::encode::utils::*;
 use crate::error::{MqttError, MqttResult};
